+++ oci/database_autonomous_data_warehouse_resource.go
@@ -85,6 +85,11 @@ func AutonomousDataWarehouseResource() *schema.Resource {
 						// Optional
 
 						// Computed
+						"all_connection_strings": {
+							Type:     schema.TypeMap,
+							Computed: true,
+							Elem:     schema.TypeString,
+						},
 						"high": {
 							Type:     schema.TypeString,
 							Computed: true,
@@ -99,6 +104,10 @@ func AutonomousDataWarehouseResource() *schema.Resource {
 						},
 					},
 				},
+			},
+			"db_version": {
+				Type:     schema.TypeString,
+				Computed: true,
 			},
 			"lifecycle_details": {
 				Type:     schema.TypeString,
@@ -364,6 +373,10 @@ func (s *AutonomousDataWarehouseResourceCrud) SetData() error {
 		s.D.Set("db_name", *s.Res.DbName)
 	}
 
+	if s.Res.DbVersion != nil {
+		s.D.Set("db_version", *s.Res.DbVersion)
+	}
+
 	if s.Res.DefinedTags != nil {
 		s.D.Set("defined_tags", definedTagsToMap(s.Res.DefinedTags))
 	}
@@ -395,6 +408,8 @@ func (s *AutonomousDataWarehouseResourceCrud) SetData() error {
 
 func AutonomousDataWarehouseConnectionStringsToMap(obj *oci_database.AutonomousDataWarehouseConnectionStrings) map[string]interface{} {
 	result := map[string]interface{}{}
+
+	result["all_connection_strings"] = obj.AllConnectionStrings
 
 	if obj.High != nil {
 		result["high"] = string(*obj.High)